+++ core/src/commands/sync/sync-start.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2023 Garden Technologies, Inc. <info@garden.io>
+ * Copyright (C) 2018-2024 Garden Technologies, Inc. <info@garden.io>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -20,6 +20,7 @@ import { createActionLog } from "../../logger/log-entry.js"
 import type { DeployAction } from "../../actions/deploy.js"
 import type { ConfigGraph } from "../../graph/config-graph.js"
 import type { Garden } from "../../index.js"
+import { styles } from "../../logger/styles.js"
 
 const syncStartArgs = {
   names: new StringsParameter({
@@ -94,6 +95,17 @@ export class SyncStartCommand extends Command<Args, Opts> {
 
   async action(params: CommandParams<Args, Opts>): Promise<CommandResult<{}>> {
     const { garden, log, args, opts } = params
+
+    if (!params.parentCommand) {
+      log.warn(
+        dedent`Behaviour of ${styles.command(
+          "sync start"
+        )} is now deprecated and will be changed in a future breaking change release.
+        Instead we recommend running ${styles.command("garden deploy --sync")} or starting syncs ${styles.italic(
+          "inside"
+        )} the dev console with either ${styles.command("deploy --sync")} or ${styles.command("sync start")}.`
+      )
+    }
 
     // We default to starting syncs for all Deploy actions
     const names = args.names || ["*"]