+++ core/src/mutagen.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2023 Garden Technologies, Inc. <info@garden.io>
+ * Copyright (C) 2018-2024 Garden Technologies, Inc. <info@garden.io>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -29,7 +29,7 @@ import { registerCleanupFunction, sleep } from "./util/util.js"
 import type { OctalPermissionMask } from "./plugins/kubernetes/types.js"
 import { styles } from "./logger/styles.js"
 import { dirname } from "node:path"
-import { makeDocsLink } from "./docs/common.js"
+import { makeDocsLinkStyled } from "./docs/common.js"
 import { syncGuideRelPath } from "./plugins/kubernetes/constants.js"
 
 const { mkdirp, pathExists } = fsExtra
@@ -331,9 +331,7 @@ function logMutagenDaemonWarning(log: Log) {
   log.warn(
     dedent`
     It looks like you've changed to a different version of the sync daemon and therefore the sync daemon needs to be restarted.
-    Please see our Troubleshooting docs for instructions on how to restart the daemon for your platform: ${styles.link(
-      makeDocsLink("guides/code-synchronization", "#restarting-sync-daemon")
-    )}}`
+    Please see our Troubleshooting docs for instructions on how to restart the daemon for your platform: ${makeDocsLinkStyled("guides/code-synchronization", "#restarting-sync-daemon")}}`
   )
 }
 
@@ -1137,7 +1135,7 @@ function formatSyncConflict(sourceDescription: string, targetDescription: string
       "one-way-replica"
     )} or ${styles.highlight("one-way-replica-reverse")} sync modes instead.
 
-    See the code synchronization guide for more details: ${styles.link(makeDocsLink(syncGuideRelPath, "#sync-modes"))}`
+    See the code synchronization guide for more details: ${makeDocsLinkStyled(syncGuideRelPath, "#sync-modes")}`
 }
 
 /**