+++ core/src/util/util.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2023 Garden Technologies, Inc. <info@garden.io>
+ * Copyright (C) 2018-2024 Garden Technologies, Inc. <info@garden.io>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -46,8 +46,7 @@ import split2 from "split2"
 import type { ExecaError, Options as ExecaOptions } from "execa"
 import { execa } from "execa"
 import corePackageJson from "../../package.json" assert { type: "json" }
-import { makeDocsLink } from "../docs/common.js"
-import { styles } from "../logger/styles.js"
+import { makeDocsLinkStyled } from "../docs/common.js"
 
 export { apply as jsonMerge } from "json-merge-patch"
 
@@ -197,9 +196,7 @@ export async function exec(cmd: string, args: string[], opts: ExecOpts = {}) {
         message: dedent`
         Received EMFILE (Too many open files) error when running ${cmd}.
 
-        This may mean there are too many files in the project, and that you need to exclude large dependency directories. Please see ${styles.link(
-          makeDocsLink("using-garden/configuration-overview", "#including-excluding-files-and-directories")
-        )} for information on how to do that.
+        This may mean there are too many files in the project, and that you need to exclude large dependency directories. Please see ${makeDocsLinkStyled("using-garden/configuration-overview", "#including-excluding-files-and-directories")} for information on how to do that.
 
         This can also be due to limits on open file descriptors being too low. Here is one guide on how to configure those limits for different platforms: https://docs.riak.com/riak/kv/latest/using/performance/open-files-limit/index.html
         `,